+++ src/cljs/snapshot/lumo/repl.cljs
@@ -17,7 +17,8 @@
             [fipp.edn :as fipp]
             [lumo.js-deps :as deps]
             [lumo.common :as common]
-            [lumo.repl-resources :refer [special-doc-map repl-special-doc-map]])
+            [lumo.repl-resources :refer [special-doc-map repl-special-doc-map]]
+            [lumo.pprint.data])
   (:import [goog.string StringBuffer]))
 
 ;; =============================================================================
@@ -152,6 +153,7 @@
         fipp.deque
         lazy-map.core
         lumo.core
+        lumo.pprint.data
         lumo.repl
         lumo.repl-resources
         lumo.js-deps
@@ -409,23 +411,24 @@
 (defn- ^:boolean could-not-eval? [msg]
   (boolean (re-find could-not-eval-regex msg)))
 
-(defn- handle-repl-error [error]
+(defn- handle-error [error]
   (let [message (ex-message error)
         cause (ex-cause error)]
-    (cond
-      (could-not-eval? message)
-      (let [message (ex-message cause)
-            {:keys [column]} (ex-data cause)
-            column-indicator-str (str (apply str
-                                        (repeat (+ (count (name @current-ns)) 3 column) " "))
-                                   "⬆")]
-        (println column-indicator-str)
-        (println message))
-      (= message "ERROR")
-      (println (str cause))
+    (binding [*print-fn* *print-err-fn*]
+      (cond
+        (could-not-eval? message)
+        (let [message (ex-message cause)
+              {:keys [column]} (ex-data cause)
+              column-indicator-str (str (apply str
+                                          (repeat (+ (count (name @current-ns)) 3 column) " "))
+                                     "⬆")]
+          (println column-indicator-str)
+          (println message))
+        (= message "ERROR")
+        (println (str cause))
 
-      :else
-      (println error))
+        :else
+        (println error)))
     (when-not (:repl? @app-opts)
       (js/$$LUMO_GLOBALS.setExitValue 1))))
 
@@ -569,7 +572,7 @@
           (cljs/eval st maybe-quoted opts
             (fn [{:keys [error value]}]
               (if error
-                (handle-repl-error error)
+                (handle-error error)
                 (let [ns-name value]
                   (if-not (symbol? ns-name)
                     (binding [*print-fn* *print-err-fn*]
@@ -580,7 +583,7 @@
                         (cljs/eval st ns-form opts
                           (fn [{:keys [error]}]
                             (if error
-                              (handle-repl-error error)
+                              (handle-error error)
                               (vreset! current-ns ns-name))))))))))))]
     (wrap-special-fns wrap-self
       {'in-ns in-ns-fn
@@ -683,8 +686,8 @@
     (= "EOF while reading string" msg)))
 
 (defn print-value [value]
-  (prn value)
-  #_(fipp/pprint value))
+  #_(prn value)
+  (lumo.pprint.data/pprint value))
 
 (defn- read-chars
   [reader]
@@ -733,14 +736,14 @@
             :js (cljs/process-macros-deps {:*compiler* st} cache nil
                   (fn [{:keys [error]}]
                     (if-not (nil? error)
-                      (handle-repl-error error)
+                      (handle-error error)
                       (cljs/process-libs-deps {:*compiler* st} cache nil
                         (fn [{:keys [error]}]
                           (if-not (nil? error)
-                            (handle-repl-error error)
+                            (handle-error error)
                             (caching-node-eval {:source source
                                                 :filename filename})))))))))
-        (handle-repl-error (ex-info (str "Could not load file " file) {}))))))
+        (handle-error (ex-info (str "Could not load file " file) {}))))))
 
 (defn- execute-text
   [source {:keys [expression? print-nil-result? filename] :as opts}]
@@ -777,9 +780,9 @@
                             (not (nil? value)))
                     (print-value value))
                   (vreset! current-ns ns))
-                (handle-repl-error error)))))))
+                (handle-error error)))))))
     (catch :default e
-      (handle-repl-error e)))
+      (handle-error e)))
   nil)
 
 (defn- execute-source
@@ -806,7 +809,7 @@
           (= "EOF" msg) ""
           (reader-eof? msg) false
           :else (do
-                  ;(handle-repl-error e)
+                  ;(handle-error e)
                   ""))))))
 
 (defn- ^:export run-main
@@ -820,7 +823,7 @@
         opts
         (fn [{:keys [ns value error] :as ret}]
           (if error
-            (handle-repl-error error)
+            (handle-error error)
             (cljs/eval-str st
               (str "(var -main)")
               nil
@@ -829,7 +832,7 @@
                 (try
                   (apply value main-args)
                   (catch :default e
-                    (handle-repl-error e)))))))))
+                    (handle-error e)))))))))
     nil))
 
 (defn ^:export get-current-ns []
@@ -848,6 +851,7 @@
                      :static-fns static-fns
                      :elide-asserts elide-asserts})
   (setup-assert! elide-asserts)
+  (set! *print-namespace-maps* repl?)
   (common/load-core-analysis-caches st repl?)
   (deps/index-upstream-foreign-libs))