+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 - Return exit code `1` on exception ([#58](https://github.com/anmonteiro/lumo/issues/58)).
 
+### Changes
+
+- Print namespaced maps at the REPL.
+
+### Bug fixes
+
+- `print` always outputs newline ([#116](https://github.com/anmonteiro/lumo/issues/116)).
+
 ## [1.3.0](https://github.com/anmonteiro/lumo/compare/1.2.0...1.3.0) (2017-04-02)
 
 ### New features